patched metallicity
Metallicity is now only written for gas particles. Line 509 was edited, but the original code is commented in case another behaviour is wanted.

Using gadget2 binary format would cause errors related to metallicity. Prof. Rubens Machado realized it was related to the existence of metallicity in particles which were not intended to have this property (bulge and disk particles, types 2 and 3).